Put an eye on it.
